fireball/utils/config.py
"""
Configuration utilities.
"""
from typing import Any, Dict, Optional
from pathlib import Path
from dotenv import load_dotenv

class Config:
    """Configuration manager."""
    
    def __init__(self, env_file: Optional[str] = None):
        """Initialize configuration."""
        if env_file:
            load_dotenv(env_file)

    def get_credentials(self, platform: str) -> Dict[str, str]:
        """Get platform credentials."""
        pass

    def get(self, key: str, default: Any = None) -> Any:
        """Get configuration value."""
        pass 