requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5
pydantic>=2.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
pandas>=2.0.0
python-docx>=1.0.0
spacy>=3.7.0
tqdm>=4.66.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 